Add generate_experiments method to PayuManager. This will basically take a list of node counts, generate layouts that match the user-supplied criteria, and create the corresponding experiments.
Currently the method still has some code that is somehow ESM1.6 specific (e.g., the branch naming scheme), but we need to have more supported models before deciding on the best way to generalise it.

Deleted the generate_experiments method, as this should not be needed anymore. If other types of generation is needed (i.e, not scaling tests), one should implement specific methods for those.